fixing Indexes

+++ db/migrate/20150121185107_fix_indexes.rb
@@ -2,18 +2,18 @@ class FixIndexes < ActiveRecord::Migration
   def change
 
     # remove long named indexes
-    remove_index :item_course_connections, [:course_id, :item_id]
-    remove_index :item_connections, [:expires_on, :item_id]
-    remove_index :item_connections, [:expires_on, :student_id]
-    remove_index :item_connections, [:student_id, :item_id, :expires_on]
+    remove_index :item_course_connections, [:course_id, :item_id] if index_exists?(:item_course_connections, [:course_id, :item_id])
+    remove_index :item_connections, [:expires_on, :item_id] if index_exists?(:item_connections, [:student_id, :item_id])
+    remove_index :item_connections, [:expires_on, :student_id] if index_exists?(:item_connections, [:student_id, :expires_on])
+    remove_index :item_connections, [:student_id, :item_id, :expires_on] if index_exists?(:item_connections, [:student_id, :item_id, :expires_on])
 
 
     # add shorter named indexes
-    add_index :item_course_connections, [:course_id, :item_id],  name: "icc_c_i_index"
+    add_index :item_course_connections, [:course_id, :item_id],  name: "icc_c_i_index" unless index_exists?(:item_course_conections, [:course_id, :item_id], name: "icc_c_i_index")
 
-    add_index :item_connections, [:expires_on, :item_id],  name: "ic_e_i_index"
-    add_index :item_connections, [:expires_on, :student_id],  name: "ic_e_s_index"
-    add_index :item_connections, [:student_id, :item_id, :expires_on],  name: "ic_s_i_e_index"
+    add_index :item_connections, [:expires_on, :item_id],  name: "ic_e_i_index" unless index_exists?(:item_connections, [:expires_on, :item_id], name: "ic_e_i_index")
+    add_index :item_connections, [:expires_on, :student_id],  name: "ic_e_s_index" unless index_exists?(:item_connections, [:expires_on, :student_id], name: "ic_e_s_index")
+    add_index :item_connections, [:student_id, :item_id, :expires_on],  name: "ic_s_i_e_index" unless index_exists?(:item_connections, [:student_id, :item_id, :expires_on], name: "ic_s_i_e_index")
 
   end
 end
